backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
boto3
openai
python-multipart
python-dotenv
httpx
Pillow
requests